requirements.txt
# Web框架
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2

# 语言模型和向量存储
langchain-openai>=0.0.1
langchain-core>=0.1.1
langchain-text-splitters>=0.0.1
langchain-milvus>=0.0.1
openai>=1.3.0
pymilvus>=2.3.0

# 数据处理和爬虫
beautifulsoup4>=4.12.2
lxml>=4.9.3
requests>=2.31.0

# 工具库
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# 数据库
pymysql>=1.1.0

# 调试工具
debugpy>=1.6.7
ipython>=8.12.0

# 添加apscheduler依赖
apscheduler>=3.10.1
